Adding jar file instrumentation and packaging
Adding to Main.java to loop over all files in a jar file. Any class
files found will be instrumented (as per a single class file) and
everything else will be copied to the new file.

Also updated the package name and added a maven repository export
to help testing.