Works with JDK9 String Concat Optimizations
With JDK9 Oracle introduced some Optimizations for concatenating
Strings. See e.g., http://openjdk.java.net/jeps/280 or
http://cr.openjdk.java.net/~shade/8085796/notes.txt. This requires us
for our instrumentation to work around this optimization slightly.

This commit solves the issue by making an Object array out of the
parameters on the stack. We then can pass the bootstrap method arguments
together with the parameter value object array to a static concatenation
method. This method replaces all placeholders with the provided
arguments and keeps track of the taints.