Added very basic toString() handling
The toString method is frequently called implicitly, e.g. when passing
objects to System.out.println. This is pretty difficult to detect.

In the future:
We could rename the toString method to e.g. $toString,
and rewrite all calls to it. This would allow to create proxies for
printing methods where we can check via reflection whether they have
our new $toString method and call that one.
This would require us to write a huge amount of proxies however..

The toString() method is not recommended as a serialization method, so
losing the taint here might not be dramatic in theory. In practice
however I'd suspect a huge amount of Java code does exactly this, so
losing taints here is kinda dramatic..