Fixing the toString issue
This isn't perfect yet, but the current idea is as follows:
For every class we instrument, we rename the toString() method to
$toString and instrument it. It thus returns a taint-aware String.
This is of course not compatible with JDK methods taking Object
instances and calling toString on them anymore.

Thus we generate a proxy toString() method, which just calls the
instrumented one. It then does the taint checking, converts it to
a regular String and returns the regular String.
In absence of proxy methods for said problematic functions, this seems
like the best solution so far.