System Test Framework can now handler .jars
To allow testing larger applications (with e.g., user interaction) this
commit adds support to test .jar files.

Every jar to test has to have a build.sh in its corresponding folder,
which just calls a build command. Additionally the folder has to be
added to the main build.sh in the tests/jar folder. This isn't really
pretty but not sure the overhead of doing it with Gradle would be worth
it...