Inner class support for Test Framework
Originally only the class file corresponding to the source entry in the
config file was instrumented. Namely for "TestString.java" only
"TestString.class" was picked.

This change adapts the Algorithm that for the example above, all classes
matching the GLOB TestString*.class are instrumented. This also picks
files like TestString$Inner.class, which are generated for inner
classes.