Added comprehensive detection of JDK classes
Initially only classes whose owner started with java were detected as
belonging to the JDK standard library. This was incomplete, as it did
not catch e.g., javax.

This change fixes this issue, as it loads the complete list of classes
belonging to the JDK standard library and checks whether the owner is
among them.