Proposed Fix for static final attributes
Under the assumption that the <clinit> constructor is always at the end
of the class file, this adds initialization code for static final String
fields in the order of occurrence to the static initializer.

This assumption has to be checked against different JDKs.

If this is correct, this saves us doing multiple passes over class
files, which makes the whole thing a lot simpler and faster.