Proper solution to static initialization.
This gets rid of the assumption that the static initialization block
will always be last in the class file.

We now only require the static initialization block to not be annotated.
I don't believe this to be legal, but did not find the relevant part in
the Java Language Specification yet.