First version of Java agent
The framework is now usable as a java agent! Call it via:
```
java -javaagent:asm_test-0.0.1-SNAPSHOT.jar StaticInitializerTest
```

The list of JDK classes is kinda broken for JDK >=9. This is relevant as
the agent attempts to transform JDK internal classes too if left
unchecked.

The script to extract the classes is broken for all JDKs supporting
modules, due to the changed format.