First attempt to fix generic/standard library issue
Some methods in the standard library return e.g., Set<String>. In the
bytecode we only see the return of Set. Every method interacting with
those Sets returns Objects directly followed by a checkcast instruction
to String.

Before this patch we just rewrote all checkcast instructions casting to
String to make them cast to taint-aware String. This breaks in this
instance as we can't cast String to TString.

This change attempts to solve this issue by keeping track of JDK methods
returning Object directly followed by a checkcast to String. If that is
detected, the checkcast instruction is replaced by a method call to
a conversion function which checks for the appropriate conversion for
the actual type. Namely conversion from String for Strings or a cast for
taint-aware Strings.