Support to pass taints through lossy methods.
Some functions lose taints due to e.g., converting the String to a
char array and manipulating the array before turning it into a String
again.

Before those would simply lose all taint information.

Now there is an option in the configuration to declare methods where the
taint information shall be propagated regardless. The taint is stored
before invocation and restored afterwards.

This currently only works for String -> String functions, which should
be the common case. If other forms of functions should be supported this
should be easy to adapt however.