Fixed runtime conversion of 2d Arrays
The code in the test case is lifted from the following call:
`org.apache.logging.log4j.core.util.datetime.FastDateParser$TimeZoneStrategy.<init>`

This did fail before, because Fontus did not detect String[][] as
something it can convert. This adds the required logic to the
ConversionUtils.

This adds an additional branch in the hot path, so might have
performance impact..