Added support for default Properties.
The Properties class is a thorn in our side since basically forever. We
have something in place that mostly works, but sometimes goes up in
a ball of flames. I found a case of the latter today.

Unlike the JDK Properties, ours did not expose the defaults field which
allows to delegate lookup to an upper Property map. This fixes the
presence and usage of the defaults field.

It's still not fully supported, as we e.g., do not write out the default
keys when `stringPropertyNames()` is called.