Conditionally add interface proxies for taint persistence
If we always load the proxies and do not use the taint persistence
driver, which might be the sensible path due to
performance/compatibility restrictions the application will break.

As the focus is not purely on GDPR tainting anymore, the addition of the
problematic proxies is hidden behind a flag has has to be passed to the
agent.

It would be nicer to dynamically detect this, but this a) might be slow
and b) is fairly complex.